+++ live-light-client/debian/logrotate.d
@@ -1,8 +1,0 @@
-/usr/share/live-light-client/*.log {
-    daily
-    rotate 100
-    missingok
-    notifempty
-    delaycompress
-    compress
-}