+++ CHANGELOG.md
@@ -7,9 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2020-04-10
+### Added
+- A schedule, you specify what days and what time the live light should be active.
+- Active and Inactive light color, as a config option.
+- A way to turn off the live light add a new endpoint (DELETE) `/color`.
+
+### Changed
+- Config format to contain new sections.
+- Debconf `config` and `templates` in `live-light-client` for new config values.
+
 ## [0.1.0] - 2020-04-04
 ### Added
 - Initial Release.
 
-[Unreleased]: https://gitlab.com/hmajid2301/markdown-mermaid-to-images/-/compare/release%2F0.1.0...master
+[Unreleased]: https://gitlab.com/hmajid2301/markdown-mermaid-to-images/-/compare/release%2F0.2.0...master
+[0.2.0]: https://gitlab.com/hmajid2301/markdown-mermaid-to-images/-/tags/release%2F0.2.0...release%2F0.1.0
 [0.1.0]: https://gitlab.com/hmajid2301/markdown-mermaid-to-images/-/tags/release%2F0.1.0